generator: java: support payload trim
Payload trim is a feature in mavlink that excludes the trailing
data of 0 bytes in the payload and thus can reduce the size of
the packet. This usually means it is possible to receive a message
from the network with less payload bytes and it should be carefully
taken care of in order to avoid unexpected results in unmarshaller.

In Java just need to make sure payload isn't accessed out of bound
will make things work.