generator: java: add mavlink2 support
mavlink2 is almost same as mavlink1 in many aspects but
differs in:

1. Added incompat_flags & compat_flags in the header
2. Message id extended to 3 bytes
3. Added signature & signing thing for security

The difference is handled in packet level and MAVLink
message is kept unchanged and this benifits transparent
protocol transition from v1.0 to v2.0 in application in
short term.

LIMITATION: the patch doesn't include the support to the
signature and signing in MAVLink2, need to be addressed
later.